% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfData}
\alias{cfData}
\title{Infant and Young Children Indicators Dataset}
\format{
A data frame with 9 columns and 18487 rows.


**Variables** | **Description**
:--- | :---
*csex* |
*calc_age_months* |
*child_bf* |
*child_eibf* |
*child_eibf_hrs* |
*child_eibf_days* |
*chhild_addbf* |
*chhild_addbf1* |
*chhild_addbf2* |
*chhild_addbf3* |
*chhild_addbf4* |
*chhild_addbf5* |
*chhild_addbf888* |
*child_bfyest* |
*note_liquid* |
*note_liquid2* |
*child_vitdrop* |
*child_ors* |
*child_water* |
*child_juice* |
*child_broth* |
*child_porridge* |
*child_bms* |
*child_bms_freq* |
*child_milk* |
*child_milk_freq* |
*child_mproduct* |
*child_mproduct_freq* |
*child_liquid* |
*note_solid_guide* |
*info_solid_food* |
*child_rice* |
*child_potatoes* |
*child_pumpkin* |
*child_beans* |
*child_leafyveg* |
*child_mango* |
*child_fruit* |
*child_organ* |
*child_beef* |
*child_fish* |
*child_insects* |
*child_eggs* |
*child_yogurt* |
*child_fat* |
*child_plam* |
*child_sweets* |
*child_condiments* |
*child_food_freq* |
}
\source{
will update later!
}
\usage{
cfData
}
\description{
This sample dataset (`bfData`) contained 18,487 observations and 9 variables.
The detailed variable's description are mentioned in the following table.
}
\examples{
# explore the first 6 observations from the dataset
head(bfData)

}
\keyword{datasets}
