% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_group_score.R
\name{fg_score}
\alias{fg_score}
\alias{fg_breastmilk}
\alias{fg_grains}
\alias{fg_pulses}
\alias{fg_milk}
\alias{fg_meat}
\alias{fg_egg}
\alias{fg_vita_fruveg}
\alias{fg_oth_fruveg}
\alias{dummy_gen}
\title{Construct the food groups consumption score (from Minimum Dietary
   Diversity 6-23 months (MDD))}
\usage{
fg_score(
  breastmilk,
  grains,
  pulses,
  dairy,
  meat,
  eggs,
  vita_fruveg,
  oth_fruveg
)

fg_breastmilk(q4, age)

fg_grains(q7b, q7d, age)

fg_pulses(q7n, age)

fg_milk(q6b, q6c, q6d, q7a, q7o, age)

fg_meat(q7i, q7j, q7k, q7m, age)

fg_egg(q7l, age)

fg_vita_fruveg(q7c, q7e, q7g, age)

fg_oth_fruveg(q7h, q7f, age)

dummy_gen(var_list = NULL)
}
\arguments{
\item{breastmilk}{child received breastfeeding or not}

\item{grains}{child consumed grains food group or not}

\item{pulses}{child consumed pulses food group or not}

\item{dairy}{child consumed diary and milk product food group or not}

\item{meat}{child consumed meat food group or not}

\item{eggs}{child consumed eggs food group or not}

\item{vita_fruveg}{child consumed vitamin A-rich fruits and vegetables or not}

\item{oth_fruveg}{child consumed fruits and vegetables or not}

\item{q4}{The binary variable which mentioned that the child was receiving
   breastfeeding in the previous day (yes = "1", no = "0").

The following variables were use for solid food group construction and it
should be code as "1" for meeting the condition (consumed in last 24 hours),
"0" for didn't (not consumed in last 24 hours), and "9" for don't know.}

\item{age}{This parameter holds the information about child age in the month
format.}

\item{q7b}{Porridge, bread, rice, noodles, and pasta.}

\item{q7d}{Plantains, white potatoes, white yams, manioc, cassava}

\item{q7n}{Beans, peas, lentils, nuts , seeds}

\item{q6b}{Infant formula}

\item{q6c}{Milk from animals, such as fresh, tinned or powdered milk}

\item{q6d}{Yogurt drinks}

\item{q7a}{Yogurt, other than yogurt drinks}

\item{q7o}{Hard or soft cheese}

\item{q7i}{Liver, kidney, heart}

\item{q7j}{Sausages, hot dogs, ham, bacon, salami, canned meat}

\item{q7k}{Any other meat, such as beef, pork, lamb, goat, chicken, duck}

\item{q7m}{Fresh fish, dried fish or shellfish}

\item{q7l}{Eggs}

\item{q7c}{Pumpkin, carrots, sweet red peppers, squash or sweet potatoes
that are yellow or orange inside}

\item{q7e}{Dark green leafy vegetables}

\item{q7g}{Ripe mangoes, ripe papayas}

\item{q7h}{Any other fruits}

\item{q7f}{Any other vegetables}

\item{var_list}{list of food consumption variables}
}
\value{
Each command produces the respective vector with "0" and "1" binary
   result. For the `food_score` function, the return result was integer value
   at continuous scale (0-8 as the IYCF food groups have 8 food groups).
The return variable's detailed explanation is mentioned below.
}
\description{
Identification of individual 6-23 months old children dietary
   practices require the diversity of food group consumption. This command
   contains the series of function to calculate the individual food group
   (8 in total) and construct the total food group consumption score required
   to calculate Minimum Dietary Diversity (MDD)
}
\examples{

# Individual Food Group Variable Construction

 df <- iycfData

# Individual Food Group Specific Function
 breastmilk <- fg_breastmilk(df$child_bfyest, df$calc_age_months)

 grains <- fg_grains(df$child_rice, df$child_potatoes, df$calc_age_months)

 pulses <- fg_pulses(df$child_beans, df$calc_age_months)

 dairy <- fg_milk(df$child_bms, df$child_milk, df$child_mproduct,
                  df$child_yogurt, df$child_yogurt, df$calc_age_months)

 meat <- fg_meat(df$child_organ, df$child_insects, df$child_beef,
                 df$child_fish, df$calc_age_months)

 eggs <- fg_egg(df$child_eggs, df$calc_age_months)

 vita_fruveg <- fg_vita_fruveg(df$child_pumpkin, df$child_leafyveg,
                               df$child_mango, df$calc_age_months)

 oth_fruveg <- fg_oth_fruveg(df$child_fruit, df$child_fruit,
                             df$calc_age_months)

# General Dummy variable generator function
 breastmilk <- dummy_gen(var_list = list(df$child_bfyest))

 grains <- dummy_gen(var_list = list(df$child_rice, df$child_potatoes))

 pulses <- dummy_gen(var_list = list(df$child_beans))

 dairy_list <- list(df$child_bms, df$child_milk,
                    df$child_mproduct, df$child_yogurt,
                    df$child_yogurt)

 dairy <- dummy_gen(var_list = dairy_list)

 meat_list <- list(df$child_organ, df$child_insects,
                   df$child_beef, df$child_fish)

 meat <- dummy_gen(var_list = meat_list)

 eggs <- dummy_gen(var_list = list(df$child_eggs))

 vita_fruveg_list <- list(df$child_pumpkin,
                          df$child_leafyveg,
                          df$child_mango)

 vita_fruveg <- dummy_gen(var_list = vita_fruveg_list)

 oth_fruveg <- dummy_gen(var_list = list(df$child_fruit, df$child_fruit))

 # Calculate Food Consumption Score
 food_score <- fg_score(breastmilk, grains, pulses, dairy, meat,
                        eggs, vita_fruveg, oth_fruveg)


}
\author{
Nicholus Tint Zaw
}
