% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foodscore.R
\name{get_foodscore}
\alias{get_foodscore}
\title{Construct the food groups consumption score (from Minimum Dietary
Diversity 6-23 months (MDD))}
\usage{
get_foodscore(
  breastmilk,
  grains,
  pulses,
  dairy,
  meat,
  eggs,
  vita_fruveg,
  oth_fruveg
)
}
\arguments{
\item{breastmilk}{child received breastfeeding or not}

\item{grains}{child consumed grains food group or not}

\item{pulses}{child consumed pulses food group or not}

\item{dairy}{child consumed diary and milk product food group or not}

\item{meat}{child consumed meat food group or not}

\item{eggs}{child consumed eggs food group or not}

\item{vita_fruveg}{child consumed vitamin A-rich fruits and vegetables or not}

\item{oth_fruveg}{child consumed fruits and vegetables or not}
}
\value{
the integer value at continuous scale (0-8 as the IYCF food groups
have 8 food groups).
}
\description{
Identification of individual 6-23 months old children who meet
the minimum dietary diversity requirement require the information about
how many food groups the child consumed yesterday (8 groups in total).
}
\examples{

# Individual Food Group Variable Construction

 df <- iycfData

# General Dummy variable generator function
 breastmilk <- round(runif(nrow(df), min = 0, max = 1), 0)
 grains <- round(runif(nrow(df), min = 0, max = 1), 0)
 pulses <- round(runif(nrow(df), min = 0, max = 1), 0)
 dairy <- round(runif(nrow(df), min = 0, max = 1), 0)
 meat <- round(runif(nrow(df), min = 0, max = 1), 0)
 eggs <- round(runif(nrow(df), min = 0, max = 1), 0)
 vita_fruveg <- round(runif(nrow(df), min = 0, max = 1), 0)
 oth_fruveg <- round(runif(nrow(df), min = 0, max = 1), 0)

 # Calculate Food Consumption Score
 food_score <- get_foodscore(breastmilk, grains, pulses, dairy, meat,
                             eggs, vita_fruveg, oth_fruveg)


}
\author{
Nicholus Tint Zaw
}
