% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mad.R, R/detect_mdd.R
\name{get_mad}
\alias{get_mad}
\alias{get_mdd}
\title{Determining Minimum Acceptable Diet}
\usage{
get_mad(age, q4, mdd, mmf, mmff)

get_mdd(food_score, age)
}
\arguments{
\item{age}{This parameter holds the information about child age in the month
format.}

\item{q4}{The binary variable which mentions that the child was receiving
breastfeeding in the previous day (yes = "1", no = "0").}

\item{mdd}{The binary variable which holds whether the child received the
minimum dietary diversity or not  (yes = "1", no = "0").}

\item{mmf}{The binary variable which indicates that the child got minimum
meal frequency ("1") or not ("0").}

\item{mmff}{This parameter is about the non-breastfed children received
minimum milk feeding frequency or not (yes = "1", no = "0").}

\item{food_score}{the integer variable with continuous scale which indicates
the number of food groups the child consumed in previous day}
}
\value{
binary variable indicate child get the minimum acceptable diet
   (mad =1 ) or not (mad = 0)

binary variables indicate child get the minimum dietary diversity
   (mdd = 1 ) or not (mdd = 0)
}
\description{
Identification of individual 6-23 months old children minimum
   acceptable diet based on their breastfeeding status

Identification of individual 6-23 months old children dietary
   practices meet the minimum requirement of dietary diversity (at least 5
   food group out of 8 food groups)
}
\examples{
  df <- iycfData
  df$mdd <- rbinom(n = nrow(df), size = 1, prob = 0.5)
  df$mmf <- rbinom(n = nrow(df), size = 1, prob = 0.5)
  df$mmff <- rbinom(n = nrow(df), size = 1, prob = 0.5)

  mad <- get_mad(df$calc_age_months, df$child_bfyest, df$mdd, df$mmf, df$mmff)



# Minimum Dietary Diversity
  df <- iycfData

  df$food_score <- round(runif(nrow(df), min = 0, max = 8), 0)

  mdd <- get_mdd(df$food_score, df$calc_age_months)


}
\author{
Nicholus Tint Zaw

Nicholus Tint Zaw
}
