% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_othercf.R
\name{get_isssf}
\alias{get_isssf}
\alias{get_eff}
\alias{get_swb}
\alias{get_ufc}
\alias{get_zvf}
\title{Construct other breastfeeding related indicators (from IYCF)}
\usage{
get_isssf(solid_food, age)

get_eff(egg_meat, age)

get_swb(sweet, age)

get_ufc(unhealthy, age)

get_zvf(vege_fruit, age)
}
\arguments{
\item{solid_food}{The binary variable which mentioned that the child was
received any type of solid foods yesterday (yes = 1 or no = 0).}

\item{age}{This parameter holds the information about child age in the month
format.}

\item{egg_meat}{The binary variable which mentioned that the child was
received any egg and/or flesh food yesterday}

\item{sweet}{The binary variable which mentioned that the child was
received any sweet beverage foods yesterday}

\item{unhealthy}{The binary variable which mentioned that the child was
received the unhealthy food consumption: sweets and instant foods}

\item{vege_fruit}{The binary variable which mentioned that the child was
received any vegetables or fruits yesterday}
}
\value{
binary variables indicate child met the respective complementary
   feeding status or not (yes = 1 or no = 0)
}
\description{
Identification of individual 0-23 months old children
   breastfeeding status
}
\examples{

df <- bfData

# Introduction of Solid, Semi-solid or soft foods (6-8 months)
# Solid food consumption previous day - yes/no
 solid <- list("child_rice", "child_potatoes", "child_pumpkin",
               "child_beans", "child_leafyveg", "child_mango", "child_fruit",
               "child_organ", "child_beef", "child_fish", "child_insects",
               "child_eggs", "child_yogurt", "child_fat", "child_plam",
               "child_sweets", "child_condiments")

 df$solid_food <- dummy_gen(df, solid)

 isssf <- get_isssf(df$solid_food, df$calc_age_months)


# Egg and/or Flesh food consumption (6-23 months)
# Egg and/or Flesh food consumption previous day - yes/no
 egg_meat <- list("child_organ", "child_beef", "child_fish", "child_insects",
                  "child_eggs")

 df$egg_meat <- dummy_gen(df, egg_meat)

 eff <- get_eff(df$egg_meat, df$calc_age_months)

# Sweet beverage consumption (6-23 months)
# Sweet beverage consumption previous day - yes/no
 sweet <- list("child_juice", "child_broth", "child_milk", "child_mproduct")

 df$sweet <- dummy_gen(df, sweet)

 swb <- get_swb(df$sweet, df$calc_age_months)

# Unhealthy food consumption (6-23 months)
# Unhealthy food consumption previous day - yes/no
 unhealthy <- list("child_sweets", "child_condiments")

 df$unhealthy <- dummy_gen(df, unhealthy)

 ufc <- get_ufc(df$unhealthy, df$calc_age_months)

# Zero vegetable or fruit consumption (6-23 months)
# Zero vegetable or fruit consumption previous day - yes/no
 vege_fruit <- list("child_pumpkin", "child_leafyveg", "child_mango",
                    "child_fruit")

 df$vege_fruit <- dummy_gen(df, vege_fruit)

 zvf <- get_zvf(df$vege_fruit, df$calc_age_months)

}
\author{
Nicholus Tint Zaw
}
