% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_otherbf.R
\name{get_evbf}
\alias{get_evbf}
\alias{get_eibf}
\alias{get_ebf2d}
\alias{get_mixmf}
\alias{get_cbf}
\alias{get_bof}
\title{Construct other breastfeeding related indicators (from IYCF)}
\usage{
get_evbf(q4, age)

get_eibf(age, q2, q2_hour)

get_ebf2d(q3, age)

get_mixmf(q4, age, q6b, q6c)

get_cbf(q4, age)

get_bof(q5, age)
}
\arguments{
\item{q4}{The binary variable which mentioned that the child was receiving
breastfeeding in the previous day (yes = "1", no = "0").}

\item{age}{This parameter holds the information about child age in the month
format.}

\item{q2}{the parameter indicate how long after the child was put to the
barest immediately after birth (0 = immediately, 1 = for hours and 2 = for
days)}

\item{q2_hour}{the integer parameter record the hour(s) after birth the child
was put to the breast}

\item{q3}{the binary variable indicating the child received anything else
beside breastmilk with the first two days after birth}

\item{q6b}{the binary variable indicates that the child got infant formula
feeding in the previous day}

\item{q6c}{the binary variable presents the child got any of the following
milk related food; Milk from animals, such as fresh, tinned or powdered
milk.}

\item{q5}{the binary variable presents the child received the bottle feeding
in the previous day}
}
\value{
binary variables indicate child met the respective breastfeeding
   status or not (yes = 1 or no = 0)
}
\description{
Identification of individual 0-23 months old children
   breastfeeding status; Ever breastfed, Early initiation of breastfeeding,
   Exclusive breastfeeding for the first two days after birth, Mixed milk
   feeding under 6 months, Continuous breastfeeding 12-23 months, and Bottle
   feeding 0-23 months
}
\examples{

 df <- iycfData

# Ever Breastfed
evbf <- get_evbf(df$child_bfyest, df$calc_age_months)

# Early Initiation of Breastfeeding
eibf <- get_eibf(df$calc_age_months, df$child_eibf, df$child_eibf_hrs)

# Exclusive Breastfeeding for the first two days after birth
df$q3 <- rbinom(n = nrow(df), size = 1, prob = 0.5)

ebf2d <- get_ebf2d(df$q3, df$calc_age_months)

# Mixed Milk Feeding Under 6 months
mixmf <- get_mixmf(df$child_bfyest, df$calc_age_months,
                   df$child_bms, df$child_milk)

# Continuous Breastfeeding 12-23 months
cbf <- get_cbf(df$child_bfyest, df$calc_age_months)

# Bottle Feeding 0-23 months
df$q5 <- rbinom(n = nrow(df), size = 1, prob = 0.5)

bof <- get_bof(df$q5, df$calc_age_months)

}
\author{
Nicholus Tint Zaw
}
