% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iycfData}
\alias{iycfData}
\title{Infant and Young Children Indicators Sample Dataset}
\format{
A data frame with 46 columns and 359 rows.\tabular{lll}{
   \strong{Variables} \tab \strong{WHO Question Number} \tab \strong{Description} \cr
   \emph{csex} \tab  \tab child sex \cr
   \emph{calc_age_months} \tab  \tab child age in month \cr
   \emph{child_bf} \tab Q1 \tab ever breastfed \cr
   \emph{child_eibf} \tab Q2 \tab first put to the breast - immediately \cr
   \emph{child_eibf_hrs} \tab Q2 \tab first put to the breast - hours \cr
   \emph{child_eibf_days} \tab Q2 \tab first put to the breast - days \cr
   \emph{bf_2days} \tab Q3 \tab given anything other than breast milk, first 2-days of child age \cr
   \emph{child_bfyest} \tab Q4 \tab breastfed yesterday during the day or at night \cr
   \emph{bf_bottle} \tab Q5 \tab drink anything from a bottle with a nipple yesterday \cr
   \emph{child_water} \tab Q6A \tab Plain water \cr
   \emph{child_bms} \tab Q6B \tab Infant formula \cr
   \emph{child_bms_freq} \tab Q6Bnum \tab Infant formula - frequency \cr
   \emph{child_milk} \tab Q6C \tab Milk from animals \cr
   \emph{child_milk_freq} \tab Q6Cnum \tab Milk from animals - frequency \cr
   \emph{child_milk_sweet} \tab Q6Cswt \tab Milk from animals - sweet or flavored type milk \cr
   \emph{child_mproduct} \tab Q6D \tab Yogurt drinks \cr
   \emph{child_mproduct_freq} \tab Q6Dnum \tab Yogurt drinks - frequency \cr
   \emph{child_mproduct_sweet} \tab Q6Dswt \tab Yogurt drinks - sweet or flavored type yogurt \cr
   \emph{child_chocolate} \tab Q6E \tab Chocolate-flavored drinks \cr
   \emph{child_juice} \tab Q6F \tab Fruit juice or fruit-flavored drinks \cr
   \emph{child_soda} \tab Q6G \tab Sodas, malt drinks, sports drinks or energy drinks \cr
   \emph{child_tea} \tab Q6H \tab Tea, coffee, or herbal drinks \cr
   \emph{child_tea_sweet} \tab Q6Hswt \tab Tea, coffee, or herbal drinks - sweetened \cr
   \emph{child_broth} \tab Q6I \tab Clear broth or clear soup \cr
   \emph{child_oth_drink} \tab Q6J \tab Any other liquids \cr
   \emph{child_oth_drink_sweet} \tab Q6Jswt \tab Other liquids - sweetened \cr
   \emph{child_yogurt} \tab Q7A \tab Yogurt, other than yogurt drinks \cr
   \emph{child_yogurt_frq} \tab Q7Anum \tab Yogurt, other than yogurt drinks - frequency \cr
   \emph{child_rice} \tab Q7B \tab Porridge, bread, rice, noodles, pasta \cr
   \emph{child_pumpkin} \tab Q7C \tab Pumpkin, carrots, sweet potatoes (yellow or orange inside) \cr
   \emph{child_potatoes} \tab Q7D \tab Plantains, white potatoes, white yams, manioc, cassava \cr
   \emph{child_leafyveg} \tab Q7E \tab Dark green leafy vegetables \cr
   \emph{child_oth_veg} \tab Q7F \tab Any other vegetables \cr
   \emph{child_mango} \tab Q7G \tab Ripe mangoes, ripe papayas \cr
   \emph{child_fruit} \tab Q7H \tab Any other fruits \cr
   \emph{child_organ} \tab Q7I \tab Liver, kidney, heart \cr
   \emph{child_processmeat} \tab Q7J \tab Sausages, hot dogs, ham, bacon, salami, canned meat \cr
   \emph{child_beef} \tab Q7K \tab Any other meat, such as beef, pork, lamb, goat, chicken, duck \cr
   \emph{child_eggs} \tab Q7L \tab Eggs \cr
   \emph{child_fish} \tab Q7M \tab Fresh fish, dried fish or shellfish \cr
   \emph{child_beans} \tab Q7N \tab Beans, peas, lentils, nuts , seeds \cr
   \emph{child_cheese} \tab Q7O \tab Hard or soft cheese \cr
   \emph{child_sweets} \tab Q7P \tab Sweet foods such as chocolates, candies, cakes and biscuits \cr
   \emph{child_snack} \tab Q7Q \tab Chips, crisps, puffs, French fries, instant noodles \cr
   \emph{child_oth_food} \tab Q7R \tab Any other solid, semi-solid or soft food \cr
   \emph{child_food_freq} \tab Q8 \tab number of any solid, semi-solid or soft foods yesterday \cr
}
}
\source{
CARE Myanmar
}
\usage{
iycfData
}
\description{
This sample demo dataset (\code{iycfData}) contained 359 observations and 46
variables. The following tables explain the detailed description of each
variable included in the sample dataset. The CARE International Myanmar
Country program provided this demo dataset, and all the Personal
Identifiable Information (PII) were excluded.
}
\examples{
# explore the first 6 observations from the dataset
head(iycfData)

}
\keyword{datasets}
